+++ Realization/backend/core/core-impl/src/main/java/eu/bcvsolutions/idm/core/model/service/impl/DefaultIdmIdentityRoleService.java
@@ -189,20 +189,14 @@ public class DefaultIdmIdentityRoleService
 							"%" + text + "%")
 					);
 		}
-		List<UUID> identities = filter.getIdentities();
+		List<UUID> identities = checkFilterSizeExceeded(IdmIdentityRoleFilter.PARAMETER_IDENTITY_ID, filter.getIdentities());
 		if (!identities.isEmpty()) {
-			if (identities.size() > 500) {
-				throw new IllegalArgumentException(String.format("Identities size exceeded, given [%s], max [%s]", identities.size(), 500));
-			}
 			predicates.add(
 					root.get(IdmIdentityRole_.identityContract).get(IdmIdentityContract_.identity).get(IdmIdentity_.id).in(identities) 
 			);
 		}
-		List<UUID> roles = filter.getRoles();
+		List<UUID> roles = checkFilterSizeExceeded(IdmIdentityRoleFilter.PARAMETER_ROLE_ID, filter.getRoles());
 		if (!roles.isEmpty()) {
-			if (roles.size() > 500) {
-				throw new IllegalArgumentException(String.format("Roles size exceeded, given [%s], max [%s]", roles.size(), 500));
-			}
 			predicates.add(root.get(IdmIdentityRole_.role).get(IdmRole_.id).in(roles));
 		}
 		List<String> roleEnvironments = filter.getRoleEnvironments();